requirements.txt
Flask
requests
python-dotenv
pyjwt[crypto]